src/model/requirements.txt
gensim
sklearn-contrib-lightning